move packages & rename
