+++ linen-modern/src/main/scala/x7c1/linen/modern/init/settings/preset/PresetsAllFragment.scala
@@ -82,5 +82,5 @@ class SubscriptionChangedUpdater(
 }
 
 trait OnChannelSubscribedListener { self =>
-  def onSubscribedChanged(event: SubscribeChangedEvent)
+  def onSubscribedChanged(event: SubscribeChangedEvent): Unit
 }
