+++ linen-modern/src/test/scala/x7c1/linen/modern/accessor/MyChannelAccessorTest.scala
@@ -6,12 +6,13 @@ import org.junit.runner.RunWith
 import org.robolectric.annotation.Config
 import org.robolectric.{RobolectricTestRunner, RuntimeEnvironment}
 import org.scalatest.junit.JUnitSuiteLike
+import x7c1.linen.modern.accessor.setting.MyChannelAccessor
 import x7c1.linen.modern.init.dev.DummyFactory
 
 
 @Config(manifest=Config.NONE)
 @RunWith(classOf[RobolectricTestRunner])
-class ChannelAccessorTest extends JUnitSuiteLike {
+class MyChannelAccessorTest extends JUnitSuiteLike {
 
   @Test
   def testFindFirstId() = {
@@ -20,10 +21,10 @@ class ChannelAccessorTest extends JUnitSuiteLike {
 
     val db = new LinenOpenHelper(context).getWritableDatabase
     val Some(accountId) = AccountAccessor.create(db) findAt 0 map (_.accountId)
-    val Some(channelId) = ChannelAccessor.create(db, accountId) findAt 0 map (_.channelId)
+    val Some(channelId) = MyChannelAccessor.create(db, accountId) findAt 0 map (_.channelId)
     assertEquals(2, channelId)
 
-    val channelId2 = ChannelAccessor.create(db, accountId = 123) findAt 0
+    val channelId2 = MyChannelAccessor.create(db, accountId = 123) findAt 0
     assertEquals(None, channelId2)
   }
 
@@ -34,7 +35,7 @@ class ChannelAccessorTest extends JUnitSuiteLike {
 
     val db = new LinenOpenHelper(context).getWritableDatabase
     val Some(accountId) = AccountAccessor.create(db) findAt 0 map (_.accountId)
-    val length = ChannelAccessor.create(db, accountId).length
+    val length = MyChannelAccessor.create(db, accountId).length
     assertEquals(2, length)
   }
 
@@ -45,13 +46,13 @@ class ChannelAccessorTest extends JUnitSuiteLike {
 
     val db = new LinenOpenHelper(context).getWritableDatabase
     val Some(accountId) = AccountAccessor.create(db) findAt 0 map (_.accountId)
-    val Some(c1) = ChannelAccessor.create(db, accountId).findAt(0)
+    val Some(c1) = MyChannelAccessor.create(db, accountId).findAt(0)
     assertEquals("sample channel name2", c1.name)
 
-    val Some(c2) = ChannelAccessor.create(db, accountId).findAt(1)
+    val Some(c2) = MyChannelAccessor.create(db, accountId).findAt(1)
     assertEquals("sample channel name1", c2.name)
 
-    val c3 = ChannelAccessor.create(db, accountId).findAt(2)
+    val c3 = MyChannelAccessor.create(db, accountId).findAt(2)
     assertEquals(None, c3)
   }
 }
