+++ linen-modern/src/test/scala/x7c1/linen/modern/accessor/SettingSourceAccessorTest.scala
@@ -7,6 +7,7 @@ import org.junit.runner.RunWith
 import org.robolectric.annotation.Config
 import org.robolectric.{RobolectricTestRunner, RuntimeEnvironment}
 import org.scalatest.junit.JUnitSuiteLike
+import x7c1.linen.modern.accessor.database.ChannelParts
 import x7c1.linen.modern.struct.{Channel, Date}
 
 @Config(manifest=Config.NONE)
